Fix rich textbox edit fields
Certain textboxes, like the Location Directions or Dance Series Description, were not appearing using the default styling configuration. The requests for the css and js were 404'ing. This is because the default settings for ckeditor referred to the skin 'moono' which is not included in the static files. The skin that is included is called 'moono-lisa'.