Simplify README text
I think a lot of the README text was quite overly verbose and had a lot
of redundant information and unnecessary qualifiers. The tone and
perspective also changed quite a lot.

I pretty much stripped down everything I could and tried to re-phrase
everything where it made sense. I think the README itself could still go
with a lot more work, but I will propose it as part of a larger set of
changes alongside a pull request.