Fix html5lib version mismatch
Both django-easy-pdf and djangocms-text-ckeditor have html5lib as one of
their subdependencies. Setting django-easy-pdf's version in the
requirements to >=0.1.0 causes the package to resolve to version 0.1.1,
and its subdependency xhtml2lib to version html5lib>=1.0b10, causing a
mismatch with djangocms-text-ckeditor required version,
html5lib>=0.999999.

The solution is to pin django-easy-pdf to version 0.1.0, which also
requires explicitly adding xhtml2pdf=0.1b1 to the requirements.

There's currently a pull request in djangocms-text-ckeditor to update
its use of html5lib, but it has yet to be merged. Once it is updated,
django-easy-pdf can be updated and the explicit xhtml2lib dependency can
be removed.