Fix access to QuerySet object in core/handlers.py
The call EmailAddress.objects.filter(...) returns a QuerySet object,
which later in the function is used directly as an EmailAddress object
and not as a QuerySet, causing an AttributeError to be thrown. The call
filter() is substituted by a get() call with the appropriate
try...except block to handle the possible exceptions raised by get().