Update multiple pieces of documentation
The primary change of this commit is to update the e-mail set-up page to
reflect newer changes in how the project is configured. However, I also
did a quick pass of some other pages, adjusting grammar or syntax
wherever improvements could be made.

sphinx still has a few build warnings based on how parts of the
documentation are written, and the files themselves are somewhat
inconsistent - some have their lines wrapped appropriately to 80
characters while others do not, so additional work is necessary.