Fixed #139, reference to content_types before migrations
The `instance_of` filter from django-polymorphic makes reference to content types.  Since the `EventAutocompleteForm` is loaded prior to initial migrations, this instance_of filter was being called before the relevant content types existed in the database.  The updated filter is functionally identical, but it does not reference content types and so it does not have the same issue.